Flagspec memory allocation overhaul.
It annoyed me that every time a text field was to be set (like longname, description, etc.)
the program called malloc (through strndup, which was problematic in its own right).
That's why i decided to introduce a memory pool inside of the flagspec; now malloc is called only
when the pool fills up (it's noticably faster for large inputs, unnoticable for small ones).
The twist is that the pool itself couldn't be a singular char vector, as reallocating it would lead
to pointer rot all over the program. Ergo, the pool's actually implemented as a vector of char vectors
(or a pool of pools, as i like to call it); when a single char vector fills up, the main vector advances
and a new one, having a greater size (growing exponentially), is allocated.

Overall changes summary:
 1. Changed memory allocation in the flagspec to a mempool.
 2. Combined several macros into one in parse.c (as a result of the alloc overhaul).
 3. Moved the flagspec and all of its functions into a separate file (flagspec.c) and header (flagspec.h)
 to make the code less cluttered.
 4. Used a macro to define flagspec_set_(identifier|longname), because they're essentially the same.